Optionally supply a list of regions to target
A regions.txt file can now be added to the lambda to specify a list of
regions it should target.

This file, if present, should list one or more regions separated by
newlines.  Any whitespace surrounding a line is ignored, as are blank
lines, and if the file is missing, empty, or only contains whitespace,
the lambda defaults to region it was started in.

(ref GitHub issue #7)