Use environment variables
Drop the handler_default_off entry point in favour of allowing the
default probability to be overridden through a "probability" environment
variable.  The previous "off" value is equivalent to a probability value
of "0" (or "0.0").

Also remove the regions.txt file in favour of a "regions" environment
variable containing a comma-separated list of region names, eg:
  ap-south-1,eu-west-1,us-east-1

Any whitespace around the names is ignored.

When the "regions" value is empty (the default), the region in which the
lambda was deployed is used instead.